fix: move paused videowrapper
