+++ apps/journeys/src/components/VideoWrapperPaused/VideoWrapperPaused.spec.tsx
