+++ apps/journeys/src/components/VideoWrapperPaused/VideoWrapperPaused.tsx
