+++ apps/journeys/src/components/VideoWrapperPaused/index.ts
